pyproject.toml
[project]
name = "mcp-calculator"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "google-api-python-client>=2.191.0",
    "google-auth-httplib2>=0.3.0",
    "google-auth-oauthlib>=1.3.0",
    "httpx>=0.28.1",
    "mcp>=1.26.0",
    "notion-client>=3.0.0",
]
